+++ src/visualize/+mag/+graphics/savePlots.m
@@ -5,15 +5,22 @@ function savePlots(figures, location, options)
         figures (1, :) matlab.ui.Figure
         location (1, 1) string {mustBeFolder} = "results"
         options.DotReplacement (1, 1) string = "_"
+        options.ColonReplacement (1, 1) string = ""
     end
 
     for f = figures
 
         if isvalid(f)
 
-            name = fullfile(location, replace(f.Name, ".", options.DotReplacement));
+            name = replace(f.Name, [".", ":"], [options.DotReplacement, options.ColonReplacement]);
+            name = fullfile(location, name);
 
-            savefig(f, name);
+            try
+                savefig(f, name);
+            catch exception
+                warning("Could not save figure ""%s"":\n%s", f.Name, exception.message);
+            end
+
             exportgraphics(f, fullfile(name + ".png"), Resolution = 300);
         end
     end