+++ resources/ReleaseNotes.md
@@ -6,6 +6,8 @@
 - Only filter one vector for mode changes, and one second for range changes
 - Hide unnecessary constants in meta data classes
 - Make sure custom events have padded y-axis limits
+- Fix issue with saving data with colon in name
+- Fix issue with saving too large figures causing an error
 - Fix issue with consecutive figures not opening maximized
 - Fix issue with saving figures in folder names with dots
 - Fix small typo in HK data loading